app/requirements.txt
flask
python-mecab-ko
base58